Resolve split package between bombe and bombe-jar
